+++ src/store/slices/theme-slice.ts
@@ -1,6 +1,6 @@
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
-import type { SystemTheme } from '../../types/themeTypes';
-import { DEFAULT_SYSTEM_THEME } from '../../constants/defaultTheme';
+import type { SystemTheme } from '../../types/theme-types';
+import { DEFAULT_SYSTEM_THEME } from '../../constants/default-theme';
 
 interface ThemeState {
   mode: 'light' | 'dark';
